+++ src/components/hooks/Restore.tsx
@@ -7,9 +7,9 @@ import { type NavigationProp, useNavigation } from '@react-navigation/core'
 import { usePromptContext } from '@src/context/Prompt'
 import { NS } from '@src/i18n'
 import { addToHistory } from '@store/latestHistoryEntries'
-import { getCounterByMintUrl, incrementCounterByMintUrl, saveSeed } from '@store/restore'
+import { saveSeed } from '@store/restore'
 import { isErr } from '@util'
-import { _setKeys, getSeedWalletByMnemonic } from '@wallet'
+import { _setKeys, getCounterByMintUrl, getSeedWalletByMnemonic, incrementCounterByMintUrl } from '@wallet'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
